pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openwrt-mcu-bridge"
version = "2.0.1"
description = "Arduino MCU Bridge Protocol for OpenWrt (Python 3.13)"
requires-python = ">=3.13"
dependencies = [
    # pyserial removed: replaced by built-in termios (mcubridge.transport.termios_serial)
    "paho-mqtt==2.1.0",      # STRICT: Required for CallbackAPIVersion.VERSION2
    "aiomqtt==2.5.0",        # STRICT: Python 3.13 compatible, validated against paho 2.1.0
    "cobs>=1.2.2",
    "psutil>=6.0.0",
    "prometheus-client>=0.21.0"
]

[tool.setuptools]
packages = ["mcubridge"]
package-dir = {"" = "openwrt-mcu-bridge"}

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --asyncio-mode=auto"
testpaths = [
    "openwrt-mcu-bridge/tests",
]
pythonpath = [
    ".",
    "openwrt-mcu-bridge"
]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pyright]
include = ["openwrt-mcu-bridge/mcubridge", "openwrt-mcu-examples-python"]
exclude = ["**/__pycache__", "**/node_modules", "**/.*", ".venv_debug", '.venv']
venv = ".venv"
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.13"
executionEnvironments = [
  { root = "openwrt-mcu-bridge/mcubridge" }
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "C90"]

[tool.ruff.lint.per-file-ignores]
# Code generators are inherently complex: exclude from mccabe checks
"tools/protocol/generate.py" = ["C901"]
# Test helper functions can have complex logic for static analysis
"openwrt-mcu-bridge/tests/test_repo_policy_gates.py" = ["C901"]

[tool.ruff.lint.mccabe]
# [QUALITY] Reduced from 50 to 25 for SIL-2 compliance and maintainability
max-complexity = 25
